services/documents/requirements.txt
jinja2>=3.1
weasyprint>=60.0
psycopg2-binary>=2.9
click>=8.1
